split to_string method
